Enable non-standard SMTP AUTH LOGIN
Turns out that the library supports this, even with some rudimentary
autodetection. The problem was that the autodetection is not available
with our queue-commands-at-once approach, and that the code in Trojita
which called this was deliberately setting AUTH PLAIN for
"compatibility", see commit c5693ba7b9fbd40e55230c37711e4cc0e3588696.

Anyway, this is a hack, but it works. Did I say that the SMTP client
should be replaced?

BUG: 342875
Change-Id: I0911ca6aa2ea98099f4b14ee103c3465e7223122