Do not call dataChanged on unsynchronized messages
If the message's UID is not known yet, we cannot store its flags or even
find out whether the command which we've just sent out actually updated
its flags.

The good thing is that the message is going to be synced, including its
flags, afterwards, so there's no actual race.

Change-Id: Ib481b036ef26f6969aa864b8f3860467fbebfae2
CCBUG: 343656