Don't stay offline during the first configuration
There was a nice bug where the network watcher intercepted the
networkError's setNetworkOffline() callout and updated the on-disk
preference to stay offline. When user pressed OK in the dialog, Trojita
did not reconnect.

This is the most elegant patch that I was able to come up with. Without
changing the NetworkWatcher, there aren't many chances of intercepting
the call to setNetworkXYZ.

Thanks to Elias Probst <mail@eliasprobst.eu> (eliasp on IRC) for
reporting this.

Change-Id: I65008fd9c0f1d928845be220f82408219bd3a860