Enable "Show message source" on attached messages
The use case behind this feature is a message which itself contains an
inline-forwarded message. It's good to be able to look at the forwarded
message without saving to disk.

I'm not sure whether addign this right into the AttachmentView's context
menu is a good thing, but then this is the menu which gets brought up by
a right mouse button when clicking on the forwarded e-mail's headers.
Maybe a context menu of the EnvelopeView would be better? Suggestions
welcome.

Change-Id: I9c6e86ea9145ebba862d57745caf12ba5a143ffd