Emit correct signals when updating flags
The generic dataChanged() signal should be emitted for the index which
describes the entire mailbox as well as for the index which corresponds
to the TreeItemMsgList -- both of these indexes have access to the
message numbers from within data(), and therefore both should be
updated.

Another signal which should be emitted is the
messageCountPossiblyChanged. The lack of this signal caused e.g. problems
in updates of the systray icon right after "mark folder read".

Change-Id: Ia008dcfcade217da8514c52ac8f479da2f148826