Don't expect [CLOSED] when mailbox gets invalidated prior to SELECT
The earlier fixes for handling of [CLOSED] were incomplete because they
failed to catch a situation where a mailbox gets replaced while the
SELECT is still pending. This happened when e.g. someone decided to open
a mailbox (SELECT on the way) right after switching to online from
offline, a switching which triggered LIST and therefore mailbox
replacement.

A fix for this is to reset the connection state when UNSELECT (or an
equivalent) completes.

Change-Id: I62c6d98ccde101dfa488a466f303c42e740c7c0f