Do not attempt to create systray when it is not supported
This is needed to work reasonably well when under Xephyr (a nested,
windowed X11 server) with no additional SW and therefore no support for
systray.

Qt documentation says that Qt is capable of dynamic creation of the
systray icons if the systray support becomes available later during
runtime. A problem with this is that there's no signal for this, and
therefore we cannot really call qApp->setQuitOnLastWindowClosed(...)
based on actual availability/unavailability of the tray.

Change-Id: I193146839e123ae437fee6091df06395fed772fb