Icons for replying and forwarding of e-mails
I'm playing with the KDE Plasma 5 desktop, and one of the changes which
it brought to my desktop is a new set of icons. I do not particularly
like the "reply" and "forward" icons because they are too similar to
each other when rendered in Breeze's default dimensions (and almost
indistinguishable on Qt5's native Fusion). The only difference between
them is the location of the tiny arrow, and its styling.

This change tries to fix this by bringing back the color styling from
Thunderbird where a reply is green while forward is violet. KMail/Oxygen
also uses violet for forwards, but replies are usually blue (but then
blue is the dominant color in that theme anyway).

Apart from colors, the arrows are made much more prominent at the cost
of their separation from the envelope at the background. My Inkscape-fu
is not terribly advanced, I have to admit.

This change simply adds these icons to Trojita's git; they are not being
used at this point.

License-wise, these are derived works from the Breeze icons [1] which is
licensed under LGPLv3+.

[1] https://github.com/NitruxSA/plasma-next-icons
CCBUG: 344962
Change-Id: I2887f3e09b076b0afe932a32919748e4687436cd