Control reporting of Trojitá's version information
Thanks to Jason A. Donenfeld <Jason@zx2c4.com> for his initial
suggestion which started this discussion. To prevent a possible
misunderstanding -- Jason wanted to have a choice to disable sending
these data altogether; his argument was trying to mask the details about
the MUA in use to make possible exploits harder.

This is not what we're doing with this patch, however. The information
conveyed through the ID command and the User-Agent header contains
valuable data for later troubleshooting. I've reported many bugs to both
client and server vendors when I was able to identify the name of the
software which was running at either side of the connection. Completely
removing these data will make this bug fixing much harder.

Furthermore, Trojitá is extremely easy to detect by e.g. the IMAP
server. We are one of the, er, more effective e-mail clients, and it
would be trivial to pattern-match the command stream which Trojitá
generates. This includes the IMAP tag identifiers, the set of supported
and utilized commands, and "innocent" stuff such as the order of MIME
headers, details of RFC2047 formatting, or format of MIME boundaries.

Based on the reasons above, I believe that providing an option for "do
not reveal that I'm using Trojitá" would be false advertising.

In future, we might change this to a tri-state combo box:
1) report trojita + its version + OS details,
2) report trojita + its version,
3) report just trojita

Change-Id: Ibc47da39cea55a5f62e5f116b1d394c30a20f8d0