Merge the "seen status" column with the generic icon decoration
Let's save the space a bit, and merge these two icons into a unified
indicator.

The QAction from the context menu is not removed. When I did so, the
mapping between actions and respective columns stopped working. I think
that this is acceptable cost. Maybe we should finally switch to custom
delegate for rendering.

Change-Id: I017e70793a10dd1309d08d0f318647e60a08ee26