Move the connectivity indicator into the toolbar
By default, Trojita shows a status bar at the bottom. While its size is
not exactly huge, maybe we could easily get away with not showing it at
all. This change works towards that goal by moving the connectivity
indicator to the main toolbar.

There's a spacer to make sure that the connectivity widget is either at
right, or all the way to the bottom.

The explicit code for widget realignment is due to
https://bugreports.qt.io/browse/QTBUG-2376 . It might not be very
visible in the default settings, but it's rather prominent when showing
a vertical toolbar with some custom widgets.

Change-Id: Ia775cfecce9d64619486b5563e57d75790060522