Remove statusbar
Now that each of the actually useful widgets got moved into toolbar, it
makes little sense to have the status bar always visible.

The status reports are shown as tooltips of the connection state
indicator, and I quite like this new thing. There's a problem that
"Mailbox opened" gets shown quite often, which means that we should fix
this state announcement a bit.

The link targets are shown as tooltips at the current mouse location. In
theory, this should work reasonably well, but for some reason it just
shows up for a very brief time on my machine :(.

Change-Id: Ibd347648eec6ee7ed02598d646e5b5edbb1770fe