Gui: Restore cursor on attachment content
Previously, the code would just override the cursor for the entire area
of attachment view. This patch effectively undoes that modification, but
in a slightly ugly way. IMHO, it would be a bit better to use a separate
containing widget for the entire "header section" of the attachment view
(the button and the associated bunch of labels), but one needs a QWidget
to set a cursor override, and we're using just nested layouts here.

Change-Id: I732e4c976191c0dedbdd721595b96b01e80fde14