Bind loss of EmbeddedWebView to plain QWidget hide
...wrapper for FindBar
The hideEvent handling in FindBar accesses the webview
to perform cleanups etc. assuming the webview is still
present what is no more the case when the "destroyed"
signal of the webview is emitted.
This causes access to invalidated memory and and crashes
or stalls.

So we bypass the FindBar::hideEvent and just hide the widget
(and empty the search as a bonus ;-)

The real cause of this problem is that QPointers are special, they are
getting invalidated from within QObject::~QObject(), not in response to
the QObject::destroyed(QObject*) signal. To make stuff interesting,
QWidget::~QWidget() duplicates some of the code from QObject::~QObject(),
so a destroyed(QObject*) gets emitted from within ~QWidget(), but that
QWeakPointer invalidation is only done from ~QObject, which runs later
on. This means that our QPointer was still valid, but accessing a
QWebView which has been long gone by that time resulted in a segfault.

Thanks to Dimitrios Glentadakis <dglent@free.fr> for reporting this, and
to Simon Hausmann for explaining the place where QObject invalidates
QWeakPointer instances.

Change-Id: I9d186ea68dc19918def08b486d329c1c4c89a80a