GUI: Fix segfault in address completion popup
This eventFilter was getting called even if m_completionReceiver was a
nullptr. Looking at the code, it seems that we were doing this pointer
manipulation in an unsafe way; we were setting m_completionReceiver to
nullptr explicitly but we were not disabling the eventFilter.

My current understanding (which might be wrong) is that this is caused
by the asynchronous nature of input event handling where Qt still sends
us key events even though our code has already closed the
m_completionPopup.

Thanks to Thomas Lübking for feedback and additional information here.

Change-Id: Ia0836dabb03868f8e575039053a12d32ff01b8db