Use the string builder and fast string concatenations
As per [1], this is something which saves a bunch of reallocations,
which is a good thing, and it needed just a ocuple of rebuilds locally
to get it working, so why not :).

Right now this is only enabled on Qt5 because that's what I use. This
could be made working on Qt4 as well, but some of the interesting bits
are only in 4.7 (or 4.8?), and our CI only covers 4.6 within the Qt4
family.

Thanks to Milian for reminding me that free optimizations are nice.

[1] http://blog.qt.io/blog/2011/06/13/string-concatenation-with-qstringbuilder/

Change-Id: I367466bd1a43427543cc2407d26e82c7cb0e36bc