softwrap in composewidget
instead of applying a static wrap, this paints a wrap
indicator after 78 chars.

Softwrapping only works reasonable with monospaced fonts
and since we also need to know the font before calculating
the wrap position in pixels, the font setting is moved into
the ComposeWidget constructor.

Since we now use a softwrap, the minimum size on the
widget is unreasoned (and using some static pixel value was
wrong itfp)

Instead a golden mean editor of 80em width is used

Change-Id: I779a0675662981aee9d02f093fd7a1be8ceb0cea