Automatic subscribe during mailbox creation
This comes handy especially when the mailbox view is configured to only
show subscribed mailboxes. I don't know whether it's better to default
to true for this new feature, though, so I'm playing it "safe".

The rework of the SubscribeUnsubscribeTask to operate on strings rather
than mailbox indexes is needed, because at the time we create this task,
that mailbox doesn't exist yet. Conceptually, it's OK to manipulate
subscription requests for non-existing mailboxes anyway, so arguably one
should probably have worked with mailbox names from the very beginning.

Change-Id: If2bb0066b7181896c2028f664eae29f05613cba6