Prevent assert crash on network reconnect
I haven't seen the reported segfault on my system even though I've been
a victim of quite frequent reconnects due to a HW glitch related to palm
rest and a bit tired HW RF kill switch. However, the backtrace in the
bug report suggests that this is the likely culprit, and the fix "makes
sense" to me.

BUG: 353314
Change-Id: Ib37825b476f47a7484004cad3b64e37876d5badb