Default to building with Qt5, not Qt4
The Qt4 is going to reach EOL from upstream at the end of 2015. I've
been running a Qt5 build for almost a year, and so have been many other
developers AFAIK. For now, you can still build with Qt4 by passing a
-DWITH_QT5=OFF to cmake.

The two things which are holding back a full switch to Qt5 are the OBS
builds (especially EL6 and missing configuration for EL7), and the MinGW
installer. Help with this is much appreciated.

When all that is done, we will be able to remove support for building
with Qt4. It also quite strongly suggests that the BB10 port won't
happen (or at least won't happen with a Qt4 toolchain), which probably
shouldn't surprise me at this point.

On the plus side, this means no longer having to worry about GCC 4.4,
being able to support C++11, getting rid of several hacks (hi, nullptr),
removing a bunch of #ifdefs, etc.

Change-Id: I6b384ee198356a1cb0feb4b62aae7264e5572705