Kill the Harmattan build
It's been more than half a year since I've performed a build of this
thing for the last time. I've since moved on and only use my N950 for
its Drive application, and the port is effectively an unfinished,
read-only thing.  There's been *one* download of a prebuilt .deb since
the 0.5 release.

It's been fun and the port served me more than once. It's also been a
nice exercise in decoupling the business logic from the GUI, and as
such, it was a success. Now it's time to move on, though.

An interesting fact is that this patch only removes 2677 lines. I'm
quite happy that the UI-specific part is such a small thing (even if it
was an incomplete thing).

Change-Id: I08c1b4989cc8ebe031cc66b603f02cf64cb13b4c