Remove the OBS imports
This is duplicating what is already tracked within the OBS' versioning
system (which sucks, of course, because it's an in-house SCM with poor
functionality compared to a standard git, but that's irrelevant here).

We cannot really cover these files by our CI, and I've been using the
OBS as a playground for testing any changes anyway.

Change-Id: I193d59de0820358bf4b67fe879a1ecdef497f4eb