Deal with registered keywords $Junk and $NotJunk
The functionality has been essentially copied from the way
\Flagged is dealt with. The patch does not (yet) provide icons
associated to the shortcuts for setting/unsetting each keyword.
(Current shortcuts: J for $Junk and Shift+J for $NotJunk.)
It does handle interaction as specified in the registration; i.e.,
when either keyword is set, the other (if present) is unset.

Change-Id: I5e8081d548a07e4f77ecdd1d2971a4c50a87991c
BUG: 354366