DRY special flag names
Special flag names (and keywords) such as \Seen and $MDNSent
and their lowercase versions are defined in multiple locations
in the source code, leading to differing lists being present.
This patch centralizes them in one location, essentially by adding
a QHash that holds a lowercase-to-canonical mapping to the FlagNames
struct. This global QHash replaces a QHash that was present in the
Model and a QSet that was used to filter out these flags in the
Message View Taglist.

Change-Id: Ia6e4ec1b10ce2404231ef9abae63ed4fcb1d5354