GUI: TagWidget improvements
Some special keywords ($MDNSent, $Submitted, $SubmitPending) do not have
a tailor-made presentation in the GUI. A TagWidget variant is added so
that these keywords can still be visualized (as gray, unremovable tags).

The function for adding new tags allowed (unintentionally?) for setting
multiple tags by space-separating them. With multiple tags, the check
for reserved flags failed. We now still allow the useful multiple tag
addition feature, but do a tag-per-tag reserved keyword check.

The TagWidget got changed to better reflect the three separate use cases
("user keyword", "system flag", "adding button"). We're now also using a
hard-coded black border because it makes more sense, IMHO. Users with
dark themes are probably screwed, but I believe that the level of
screwedness is a at most the same as previously.

Now that the code was rewritten to a large extent, let's change the
license to match the rest of Trojita. If this is a derived work, the
change is still OK because the original license explicitly allows such a
change.

Change-Id: Ia8f374bd45201b1d65b5a03300794550d3f90a43