GUI: Scrolling within one message by space/backspace
This functionality was already there for situations where the focus
remained in the MsgListWidget. This patch makes it possible to continue
scrolling even once the focus shifts to the message in question.

As a bonus, we're using a pretty animation for the actual scroll.

Thanks to Jan Pobříslo <ccx@webprojekty.cz> (CcxCZ on IRC) for
reporting, and to Thomas for pointing out that mixing "eights of a
degree" and "pixels" is a recipe for disaster.

Change-Id: I0e3deb230f401c46a8d16bdd34176f25f1a1fca0