Remove redundant deleted mail icon in message list
In the message list, deleted mails are shown with a red cross icon
and black strikeout text. The red cross icon replaces the icon used
to inform about other message properties (replied, forwarded,...).
Moreover, the icon is redundant due to the strikeout effect.

This patch removes the icon, i.e., an icon showing the message
state is shown instead.

Change-Id: I71e69ae39bbe36de7157c5ac051b0f9da5a2849d