Simplify monospace font selection
I'm not interested in maintaining this code now that Qt actually
provides an equivalent (thanks to Thomas for letting me know). It
appears to cover both KDE and all GTK-based DEs that we cover right now,
so it's a net improvement.

This function is a good candidate for removal when we kill the Qt4 port.

Change-Id: I1af16de607982e56a4cb4586704cf3f5c982ffe3