Enable reusing IMAP credentials for SMTP
I suspect that it's quite common to have the same credentials being used
for both IMAP and SMTP in certain environments -- and that's not just
passwords, but also the login name.

Given that there's no session-wide storage/cache of SMTP passwords
(mainly because there's no equivalent of an ImapAccess for SMTP, yet),
let's make it possible to specify that these two sets of credentials are
supposed to be kept in-sync. Doing this explicitly is IMHO better than
checking via some ad-hoc measures ("were the old logins and passwords
the same? If so, let's update both now").

Change-Id: Ic9e96e4a9f83bbe7a1a701b03dc80b6f10e13e04