Fix treatment of message size display for < 100 B case
The patch merged from Change 532 was flawed for the case
where a message has a size strictly smaller than 100 bytes:
It would give a decimal to get to three significant digits
also in this case, e.g., display 14.0 B, which is nonsense.
This is fixed in this patch.

Also fixed is the fact that zero bytes messages were not
treated separately: a logarithm of 0 would have been
requested.

Moreover, the efficiency is increased by making some local
variables static.

Change-Id: If3319aba83366be99b2efee54138011d002234e1