do not enforce a minimum size
I don't even understand why that restriction is there - eg. it would
make it impossible to use the widescreen mode on a 1024x netbook or
similar, but iffffff there's requirement for an artificial minimum size
or clever layout choice for dynamic workspace geometries - it needs to
be smarter ;-)

So here's the challenge to just remove this.

Change-Id: I01912e53b39bda8bf312d36359c3f4dda6f74549