GUI: Add an action for closing of standalone message windows
Unfortunately, one cannot use the GUI to assign for example "Esc" for
this particular shortcut because the GUI settings dialog intercepts that
one, and closes itself. However, assigning some other key for this
action and editing the config file afterwards works.

I've also double-checked that the signal-slots are correct, and that
only one popup window is closed when pressing that shortcut.

This is an alternative to https://gerrit.vesnicky.cesnet.cz/r/538 .

Change-Id: I7479be102736f071321fb7132a7f688f95c4bd7c
CCBUG: 357679