Disable KDE plugins by default
This is in preparation of the v0.6 release. Because these plugins have
very uncertain future (nobody ported them to Qt5, essentially, and Pali
says he might not continue his maintenance in future), let's make sure
that a user has to do something explicit in order to get these plugins.
Otherwise it might come as a nasty surprise that these plugins are gone
in the next version.

Change-Id: I5fcc7b78520095b05cccf12d2f6888f688cc5df1