GUI: refactor custom action for WebView's context menu
There's really no need for the MessageView to poke within the
SimplePartWidget's internal stuff and for testing its hit contents; this
should be done by the SimplePartWidget alone.

Also reorder the resulting actions in a (subjectively) better way --
context-specific actions at the top, with the global ones at the bottom.

Change-Id: If2cec5381ad1f93fb3f16f188963d9f293e94689