NSIS: Replace Qt4 with Qt5
At the core of this change is using a Python wrapper around `objdump`
which determines which DLL libraries are needed. I would love to use
`windeployqt`, but the relevant code paths are not implemented for MinGW
builds.

The code is a bit ugly at times, but it works -- it produced an
installer which installed Trojita on Windows Server 2012 R2, at least.

A key to a successful build on EL7 with EPEL is this invocation:

 cmake/bin/cmake \
   -DCMAKE_BUILD_TYPE=Release \
   -DCMAKE_SKIP_RPATH:BOOL=ON \
   -DCMAKE_TOOLCHAIN_FILE=/usr/share/mingw/toolchain-mingw32.cmake \
   -DWITH_TESTS=OFF \
   -DWITH_NSIS=ON \
   -DOpenSSL_SSL_LOC:PATH=/usr/i686-w64-mingw32/sys-root/mingw/bin/libssl-10.dll \
   -DOpenSSL_Crypto_LOC:PATH=/usr/i686-w64-mingw32/sys-root/mingw/bin/libcrypto-10.dll \
   ..

When the CI systems are upgraded to CMake 3.4, we can get rid of that
manual path passing for the OpenSSL DLLs, and just require a new enough
CMake.

Change-Id: I8c94266ebf50fc87fab589aecc3b2cf8b17f1636