Switching to the new signal-slot syntax
Of course, this is a new shiny C++11 toy, so I simply have to use it. It
might be ugly at times, but it provides compile-time checking, yay!.

There are still some cases where we still use the old SIGNAL/SLOT
macros.

I also took the liberty to introduce some small refactoring where it
made sense (and where it was too obvious to miss). This means that
there's a bunch of small lambdas instead of trivial slots.

Change-Id: I17157e2ae4c2d724a1d754aeaab420de404441ea