clazy: fix-fromLatin1_fromUtf8-allocations plus manual fixes
Unfortunately, clazy and `QTest::addRow() << ""` doesn't really play
well together.

Then we have our funny Unicode strings in the test suite, etc. At this
point, I start to think if this really makes sense, just in order to
nuke some warnings -- maybe it would be easier to leave the test suite
as-is, and only pay attention to the clazy bits for actual sources?

But on the other hand, there's definitely some value in having a
warning-free build, and I've paid that time investment for this
particular check now, anyway.

Change-Id: I03ffce14ab39eea24f3359abe113ad87303271b1