Use a real container class
We don't need a QList, it's overly inefficient. Now, we could have used
QVector if our TextInfo had a default ctor, but it doesn't have one, so
QVector is out of the game. Let's go with STL, and wow, I just found the
emplace_back which looks like something so C++11ish that I just have to
use it.

This all started as an attempt to workaround an unrelated bug within
clazy's use of clang, but this is a nice outcome, too.

The route towards this patch was "ok, clazy says QList sucks, let me
re-read why exactly" -> "ah, that's a good point, let's use QVector" ->
"hmm, no default ctor, OK, why don't we use STL in the first place?" ->
"wow, emplace_back, nice" -> "OK, let's do it properly, as Thomas
suggested".

Change-Id: Ib842d7307f53d6318c6ad58a78871234e7a35e29