Fix segfault at exit due to QSvgIconEngine being unloaded too early
On my build of Qt 5.6, the QSvgIconEngine is provided by a plugin, and
this plugin is unloaded "too early" during the application's shutdown,
earlier than the iconDict is cleared. This results in ~QIcon trying to
access an QIconEngine which is already gone, and the end result is a
spectacular crash:

 #0  ~QIconPrivate (this=0x5555557a0fb0, __in_chrg=<optimized out>) at /var/tmp/portage/dev-qt/qtgui-5.6.9999/work/qtgui-5.6.9999/src/gui/image/qicon_p.h:64
 #1  QIcon::~QIcon (this=<optimized out>, __in_chrg=<optimized out>) at /var/tmp/portage/dev-qt/qtgui-5.6.9999/work/qtgui-5.6.9999/src/gui/image/qicon.cpp:680
 #2  0x0000555555558327 in QMapNodeBase::callDestructorIfNecessary<QIcon> (t=...) at /usr/include/qt5/QtCore/qmap.h:101
 #3  0x00005555555580ed in QMapNode<QString, QIcon>::destroySubTree (this=0x5555557a0550) at /usr/include/qt5/QtCore/qmap.h:126
 #4  0x0000555555557c2e in QMapData<QString, QIcon>::destroy (this=0x555555795030) at /usr/include/qt5/QtCore/qmap.h:244
 #5  0x000055555555772f in QMap<QString, QIcon>::~QMap (this=0x55555575a038 <getIcon(QString const&)::iconCache>, __in_chrg=<optimized out>)
     at /usr/include/qt5/QtCore/qmap.h:335
 #6  0x00007ffff5c145e0 in __run_exit_handlers (status=0, listp=0x7ffff5f80618 <__exit_funcs>, run_list_atexit=run_list_atexit@entry=true) at exit.c:82
 #7  0x00007ffff5c14648 in __GI_exit (status=<optimized out>) at exit.c:104
 #8  0x00007ffff5bfd9eb in __libc_start_main (main=0x555555555bd0 <main(int, char**)>, argc=1, argv=0x7fffffffd8d8, init=<optimized out>,
     fini=<optimized out>, rtld_fini=<optimized out>, stack_end=0x7fffffffd8c8) at libc-start.c:331
 #9  0x0000555555555ac9 in _start ()

This is very similar to what was reported against Qt 5.0 [1]. However,
it seems that Qt 5.6 (or some later version) might actually get a fix in
terms of never unloading the plugins [2]; at that point this workaround
won't be needed anymore. Reported upstream according to their
preferences.

BTW, I was surprised that I wasn't able to connect the
QCoreApplication::aboutToQuit to Cache::clear due to the
QtPrivateSignal.

[1] https://bugreports.qt.io/browse/QTBUG-27005
[2] https://codereview.qt-project.org/140750/

Upstream-Bug: https://bugreports.qt.io/browse/QTBUG-50829
Change-Id: Ic740cc5fc15c2bd8a70bf4921e63061ec24f5be4