WIP: Implement OpenPGP signature verification
This is a big commit because it mixes together the underlying bits with
an UI overhaul. The newly added support for encryption necessitated some
changes, and in the end there aren't that many differences between
signed and encrypted messages, so it made sense to mix them together
(IMHO).

The QCA leaves something to be desired, unfortunately -- for example, it
is not possible to fetch keys from keyservers, the key validation
appears to be unimplemented, and accessing the PGP key info such as the
identity name is super-fragile and inefficient (one has to fetch
*everything*, for example).

But it's a reasonable proof-of-concept, IMHO.

- FIXME: more tests

Change-Id: Ia05eb62f83eccff0794f1922b523faf63f17fa7d