Support long-lived PartReplacer instances (2/2)
This is a bit hackish; a good end result would be moving the crypto code
into a proper plugin, but that's non-trivial at the moment because the
libCryptography wants to link to all sorts of Trojita-internal modules
(which means that we might have to introduce a proper .so for these,
eventually).

So at least move the PartReplacers into the PluginManager, even though
the PluginManager itself cannot instantiate them at the moment. But they
(PartReplacer instances) are long-lived now.

Change-Id: I40924b82a63665cdbf6a81989dfda826d241ed8b