Composer: do not quote RFC2231 parameters
I assumed that this was allowed due to the general MIME header syntax
rules, but this serialization apparently doesn't match the ABNF grammar,
and there are several bugreports in competing MUAs and e-mail libraries
[1] [2] about quotes not being conformant.

Motto of the day: "Even PHP is better than Trojitá". I'm so ashamed :(.
Thanks to msjasinski@gmail.com for reporting this.

[1] https://roundup.kolab.org/issue1562
[2] https://pear.php.net/bugs/bug.php?id=16706

BUG: 359656
Change-Id: I2d26f4fd862e78d2fffb2ecc61c75c5f5e241c01