Require C++11's native alignof
We are requiring full support for C++11 for the next release. The GCC
ships with it since 4.6, so it's not a problem for us. The MSVC is a bit
tricky, but apparently there's support for it in the 2013 CTP version
[1] (whatever it is), and our __alignof workaround apparently doesn't
work anyway because on MSVC it fails on abstract classes [2]. Let's stop
pretending that it works. If you're affected, please get a usable C++11
compiler, it's 2016 already.

[1] http://herbsutter.com/2013/11/18/visual-c-compiler-november-2013-ctp/
[2] http://thread.gmane.org/gmane.mail.trojita.general/1331

Change-Id: Ic005a0a2966293302956f0fa29a7ab7c3930f9b9