hint offline state and net error in systray
mail count text turns red whenever trojita is offline
and on network errors, the systray yells a message

this also catches a possible nullptr deref in handleTrayIconChange
and I took the liberty to revisit the icon overlay painting code ;-)

jkt: Do not use red-text-on-bnlue-background, that is visually too
annoying to me. When painting in red, change the background color to
gray. It would be nice to show a red-tinted icon when no text is to be
shown, but my QPainter-composition-fu is not high enough to do that.

Change-Id: Ib1a299cfcc61ad860d3e834deaa5441bd34516fc
CCBUG: 342940