Use profile name as account name to avoid overwriting passwords
When storing the password using the QtKeyChain plugin, the lack of
an account name (also called account id in places) results in the
IMAP and SMTP password being overwritten in case multiple profiles
are used. This patch uses the profile name for the account name,
so that this does not happen.

Change-Id: I3a70a4865df2c016d67c15f0f1a28aedddab071c