Use Qt5's features for QLineEdit actions
Our LineEdit started as a class offering a clear button at the right
edge even on styles where this is not available. A feature like this
however got added to Qt in version 5.2, so it makes sense to start using
it instead.

This change makes it possible to deprecate the QSS usage, and therefore
fixes the style issues on (some) dark themes or dynamic palette changes.

The Qt's features do not work on QToolButton level, but on a QAction
level. This means that we have to open the QMenu manually.

The copyright info change reflects the fact that the history code was
contributed by Thomas in 2013. I'm leaving Glad in that file because he
still started doing this button UI thingy.

Change-Id: I818da0646a0fe6ed6c346c81e70ac6d2ffa0c9fd
CCBUG: 360020