Use synthetic CSS for specifying colors of the "web" content
The usual story -- if the current color theme is dark and the QWebView
is showing an HTML message, Bad Stuff™ happens quite easily.

By going through a custom CSS stylesheet, we achieve more or less
consistent results as far as I can see -- including various HTML e-mails
that I have in my test collection. There are of course also e-mails
where it goes horribly wrong, so the real question is "should we try to
emulate dark theme on HTML content, or should we always use white
background in order to "better accommodate all content"?

Thomas was kind enough to add a heuristic which tries to arrive to font
colors which are usable on both dark and bright backgrounds, but as
usually -- it's just a heuristic, and it will break on certain inputs.

The alternative is to always and unconditionally force black-on-white. I
could do that pretty easily.

Change-Id: Iabdf9b0d97448b070fa7aa1ca4302cc6b2c8b77f
CCBUG: 360020